pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codesoul"]

[project]
name = "codesoul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastembed>=0.7.4",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "textual>=7.4.0",
    "typer>=0.21.1",
    "watchdog>=6.0.0",
]

[project.scripts]
codesoul = "codesoul.app:cli"

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
]
